refactor(frontend): change name of file
